Uncompress and merge 24 hours of 100 Hz small waveform from 13 files in 8 seconds

+++ netpod/src/lib.rs
@@ -134,7 +134,7 @@ pub struct ChannelConfig {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum Shape {
     Scalar,
-    Wave(usize),
+    Wave(u32),
 }
 
 pub mod timeunits {
